src/asim/scripts/requirements.txt
altair
asv
azure-identity
azure-storage-blob
black>=22.0,<23
bump2version
coveralls
cytoolz==0.12.*
dask==2023.11.*
descartes
filelock
fsspec
geopandas
gh
gitpython
ipykernel
isort
jupyterlab
matplotlib
multimethod==1.9
myst-parser
nbconvert
nbformat
nbmake
numba==0.57.*
numexpr
numpy==1.24.*
numpydoc
openmatrix==0.3.*
orca==1.8
ortools==9.7.2996
pandas==2.2.*
pandana==0.7
pandera>=0.15,<0.18.1
platformdirs==3.2.*
pre-commit
psutil==5.9.*
pyarrow==11.*
pycodestyle
pydantic==2.6.*
pydata-sphinx-theme
pyinstrument==4.4
pypyr==5.8.*
pytest==7.2.*
pytest-cov
pytest-regressions
pyyaml==6.*
requests==2.28.*
rich==13.3.*
ruamel.yaml
ruff
setuptools_scm
scikit-learn==1.2.*
sharrow>=2.9.1
simwrapper>1.7
snakeviz
sparse
sphinx==6.1.*
sphinx_rtd_theme==1.2.*
sphinx-argparse==0.4.*
tables>=3.9
xarray==2025.01.*
xmle
wring
zarr>=2,<3
zstandard
activitysim==1.4.0
